+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master (unreleased)
 
 - [#342](https://github.com/clojure-emacs/orchard/pull/342): Inspector: add hexdump view mode.
+- [#343](https://github.com/clojure-emacs/orchard/pull/343): Inspector: rework view-mode toggling.
 
 ## 0.34.3 (2025-04-28)